src/backend/transcription_service/requirements.txt
faster-whisper==1.0.3
openai-whisper==20231117
numpy==1.24.3
torch==2.1.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
huggingface_hub==0.19.4
